runtime: move member to class
