+++ cmd/scheduler/src/dispatch_conv.rs
@@ -30,13 +30,14 @@ pub fn discord_event_to_dispatch(evt: DispatchEvent) -> Option<DiscordDispatchEv
         DispatchEvent::MemberAdd(m) => Some(DiscordDispatchEvent {
             name: "MEMBER_ADD",
             guild_id: m.guild_id,
-            data: serde_json::to_value(&runtime_models::discord::member::Member::from(m.0))
+            data: serde_json::to_value(&runtime_models::internal::member::Member::from(m.0))
                 .unwrap(),
         }),
         DispatchEvent::MemberUpdate(m) => Some(DiscordDispatchEvent {
             name: "MEMBER_UPDATE",
             guild_id: m.guild_id,
-            data: serde_json::to_value(&runtime_models::discord::member::Member::from(*m)).unwrap(),
+            data: serde_json::to_value(&runtime_models::internal::member::Member::from(*m))
+                .unwrap(),
         }),
         DispatchEvent::MemberRemove(m) => Some(DiscordDispatchEvent {
             name: "MEMBER_REMOVE",
