+++ components/runtime/src/ts/discord/index.ts
@@ -1,4 +1,5 @@
 export * from './dapi';
 export * from './interaction';
 export * from './message';
+export * from './member';
 export * from '../generated/discord/index';
